# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/cases": {
            "get": {
                "description": "Endpoint ini mengambil seluruh data kasus dengan dukungan paginasi dan pencarian judul. Cocok untuk monitoring dan analisis kasus. Asumsi: paginasi default 10 item/halaman.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cases"
                ],
                "summary": "Mendapatkan seluruh data kasus",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nomor halaman, default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah item per halaman, default 10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pencarian berdasarkan judul kasus",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan daftar kasus beserta metadata paginasi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/cases/count": {
            "get": {
                "description": "Endpoint ini mengambil jumlah total kasus yang terdaftar dalam sistem. Cocok untuk statistik dan monitoring. Asumsi: data konsisten.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cases"
                ],
                "summary": "Mendapatkan jumlah kasus",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan jumlah kasus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/cases/create": {
            "post": {
                "description": "Endpoint ini bertanggung jawab untuk pembuatan kasus baru, termasuk validasi, transformasi input, dan penanganan error. Proses ini krusial dalam workflow investigasi. Asumsi: data valid, judul unik per lokasi dan waktu.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cases"
                ],
                "summary": "Membuat entitas kasus baru",
                "parameters": [
                    {
                        "description": "Payload kasus dalam format JSON.\n\nField 'status' hanya dapat bernilai salah satu dari: \n- 'Open': Menandakan kasus baru dibuka dan seluruh proses investigasi masih dalam tahap inisiasi. Status ini merepresentasikan fase awal dalam siklus hidup kasus, di mana pengumpulan bukti dan identifikasi pihak terkait menjadi prioritas utama. \n- 'In Progress': Mengindikasikan bahwa kasus sedang dalam proses investigasi aktif. Pada tahap ini, seluruh sumber daya dialokasikan untuk analisis, pemeriksaan saksi, dan pengembangan hipotesis. Status ini menuntut adanya pembaruan berkala dan monitoring ketat terhadap perkembangan kasus. \n- 'Closed': Menunjukkan bahwa seluruh proses investigasi telah selesai, baik karena kasus telah terpecahkan, tidak ditemukan cukup bukti, atau alasan administratif lainnya. Status ini menandai berakhirnya siklus investigasi dan seluruh data kasus diarsipkan untuk keperluan audit dan pembelajaran institusional.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cases_request.Cases_Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Kasus berhasil dibuat; response berisi data kasus baru",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Permintaan tidak valid; data input salah",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/cases/delete/{id}": {
            "delete": {
                "description": "Endpoint ini menghapus data kasus berdasarkan ID. Proses ini tidak dapat dibatalkan dan berdampak pada data terkait. Asumsi: ID valid dan eksis.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cases"
                ],
                "summary": "Menghapus data kasus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID kasus yang akan dihapus",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus kasus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/cases/latest": {
            "get": {
                "description": "Endpoint ini mengambil seluruh data kasus dengan dukungan paginasi dan pencarian judul. Cocok untuk monitoring dan analisis kasus.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cases"
                ],
                "summary": "Mendapatkan seluruh data kasus yang terakhir diperbarui",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan daftar kasus terbaru dengan maksimal 5 data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/cases/update/{id}": {
            "put": {
                "description": "Endpoint ini memperbarui data kasus berdasarkan ID, mendukung perubahan seluruh atribut utama. Validasi dan error handling dilakukan secara ketat. Asumsi: ID valid, judul tetap unik.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cases"
                ],
                "summary": "Memperbarui data kasus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID kasus yang akan diperbarui",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload kasus dalam format JSON.\n\nField 'status' hanya dapat bernilai salah satu dari: \n- 'Open': Menandakan kasus baru dibuka dan seluruh proses investigasi masih dalam tahap inisiasi. Status ini merepresentasikan fase awal dalam siklus hidup kasus, di mana pengumpulan bukti dan identifikasi pihak terkait menjadi prioritas utama. \n- 'In Progress': Mengindikasikan bahwa kasus sedang dalam proses investigasi aktif. Pada tahap ini, seluruh sumber daya dialokasikan untuk analisis, pemeriksaan saksi, dan pengembangan hipotesis. Status ini menuntut adanya pembaruan berkala dan monitoring ketat terhadap perkembangan kasus. \n- 'Closed': Menunjukkan bahwa seluruh proses investigasi telah selesai, baik karena kasus telah terpecahkan, tidak ditemukan cukup bukti, atau alasan administratif lainnya. Status ini menandai berakhirnya siklus investigasi dan seluruh data kasus diarsipkan untuk keperluan audit dan pembelajaran institusional.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cases_request.Cases_Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil memperbarui data kasus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Permintaan tidak valid; data input salah",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/cases/{id}": {
            "get": {
                "description": "Endpoint ini mengambil data kasus spesifik berdasarkan UUID. Cocok untuk detail view atau proses investigasi lanjutan. Asumsi: ID valid dan eksis.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cases"
                ],
                "summary": "Mendapatkan kasus berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID kasus yang dicari",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data kasus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cases_request.Cases_Request": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "case_description": {
                    "type": "string"
                },
                "case_title": {
                    "type": "string"
                },
                "detective_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "incident_date": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "Open",
                        "In Progress",
                        "Closed"
                    ]
                }
            }
        },
...
    }
}